requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.3
gtts
Pillow
python-dotenv
watchdog